perf(MyData): 尝试优化MyData页面在大量刷新站点时的卡顿
